+++ Unity/Assets/Actions/NeuroActionHandler.cs
@@ -3,14 +3,16 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using NeuroSdk.Il2Cpp;
 using NeuroSdk.Messages.Outgoing;
-using NeuroSdk.Utilities.Il2Cpp;
 using NeuroSdk.Websocket;
 using UnityEngine;
 
 namespace NeuroSdk.Actions
 {
+#pragma warning disable CS0618 // Type or member is obsolete
     [RegisterInIl2Cpp]
+#pragma warning restore CS0618 // Type or member is obsolete
     public sealed class NeuroActionHandler : MonoBehaviour
     {
         private static List<INeuroAction> _currentlyRegisteredActions = new();
