+++ Unity/Assets/Internal/_AssemblyInfo.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 52f2cd6b8525483ea16012c43ea325db
+timeCreated: 1738649834
