--- Unity/Assets/Utilities/Il2Cpp.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a35261b33e1e447eaff45955ad98598f
-timeCreated: 1738526806
