--- Unity/Assets/Utilities/Il2Cpp/RegisterInIl2CppAttribute.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3c558b6a872b473a9c6ff983b4c446f4
-timeCreated: 1738526823
