+++ Unity/Assets/Websocket/ExecutionResult.cs
@@ -1,5 +1,7 @@
 ﻿#nullable enable
 
+using NeuroSdk.Internal;
+
 namespace NeuroSdk.Websocket
 {
     public sealed class ExecutionResult
