+++ Unity/NuGet/NeuroSdk.csproj
@@ -24,11 +24,15 @@
         <PackageReference Include="UnityEngine.Modules" Version="2019.4.24" />
         <Compile Include="..\Assets\**\*.cs" />
         <Compile Remove="..\Assets\Examples\*.cs" />
+        <Compile Update="..\Assets\Utilities\RegisterInIl2CppAttribute.cs">
+          <Link>Utilities\RegisterInIl2CppAttribute.cs</Link>
+        </Compile>
     </ItemGroup>
 
     <ItemGroup>
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
         <Compile Remove="Example\**\*" />
+        <None Remove="Example\**\*" />
         <Folder Remove="Example\**\*" />
     </ItemGroup>
 </Project>
